skills/woocommerce-webhooks/examples/fastapi/requirements.txt
fastapi>=0.128.1
python-dotenv>=1.0.0
pytest>=9.0.2
httpx>=0.28.1